Fixed bug when creating, updating and deleting product

+++ src/main/java/com/isep/acme/rabbit/RMQListener.java
@@ -24,18 +24,28 @@ public class RMQListener {
     public void listener(Message message){
         final String action= message.getMessageProperties().getHeader("action");
         final Product product = (Product) messageConverter.fromMessage(message);
+        System.out.println("Received Product Message " + product);
+        final Optional<Product> productToAction = repository.findBySku(product.getSku());
         if(product != null){
             switch(action) {
                 case Constants.CREATED_PRODUCT_HEADER:
-                    repository.save(product);
+                    if(productToAction.isEmpty()){
+                        repository.save(product);
+                        System.out.println("Product Added " + product);
+                    }
                     break;
                 case Constants.UPDATED_PRODUCT_HEADER:
-                    final Optional<Product> productExists = repository.findBySku(product.getSku());
-                    productExists.get().updateProduct(product);
-                    repository.save(productExists.get());
+                    if(!productToAction.isEmpty()){
+                        productToAction.get().updateProduct(product);
+                        repository.save(productToAction.get());
+                        System.out.println("Product Updated " + productToAction.get());
+                    }
                     break;
                 case Constants.DELETED_PRODUCT_HEADER:
-                    repository.deleteBySku(product.getSku());
+                    if(!productToAction.isEmpty()){
+                        repository.deleteBySku(product.getSku());
+                        System.out.println("Product Deleted " + productToAction.get());
+                    }
                     break;
                 default:
                     break;
